Updated files structure.
