+++ packages/uniforms-material/package.json
@@ -1,7 +1,7 @@
 {
   "name": "uniforms-material",
-  "version": "1.7.0",
-  "main": "lib/index.js",
+  "version": "1.8.1",
+  "main": "index.js",
   "jsnext:main": "src/index.js",
   "description": "Material UI components for uniforms.",
   "repository": "https://github.com/vazco/uniforms/tree/master/packages/uniforms-material",
@@ -20,53 +20,52 @@
     "validation"
   ],
   "files": [
-    "lib/",
+    "*.js",
     "src/"
   ],
   "scripts": {
-    "build": "babel --out-dir lib src",
-    "clean": "rimraf coverage lib .nyc_output",
+    "build": "babel --out-dir . src",
+    "clean": "rimraf *.js coverage .nyc_output",
     "cover": "cross-env NODE_ENV=cover nyc npm test",
     "lint": "eslint src test",
     "prepublish": "npm run build",
     "pretest": "npm run lint",
     "reset": "rimraf node_modules",
-    "test": "babel-node node_modules/.bin/_mocha --require test/index.js --recursive test --reporter list"
+    "test": "babel-node node_modules/.bin/_mocha --require test/index.js test --reporter list"
   },
   "devDependencies": {
-    "babel-cli": "^6.18.0",
-    "babel-eslint": "^7.1.1",
-    "babel-plugin-istanbul": "^3.0.0",
-    "babel-plugin-transform-object-assign": "^6.8.0",
-    "babel-plugin-transform-react-inline-elements": "^6.8.0",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-es2015": "^6.18.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "babel-runtime": "^6.18.0",
-    "chai": "^4.0.0-canary.1",
-    "cross-env": "^3.1.3",
-    "enzyme": "^2.6.0",
-    "eslint": "^3.11.1",
+    "babel-cli": "6.18.0",
+    "babel-eslint": "7.1.1",
+    "babel-plugin-istanbul": "3.0.0",
+    "babel-plugin-transform-object-assign": "6.8.0",
+    "babel-plugin-transform-react-inline-elements": "6.8.0",
+    "babel-plugin-transform-runtime": "6.15.0",
+    "babel-preset-es2015": "6.18.0",
+    "babel-preset-react": "6.16.0",
+    "babel-preset-stage-0": "6.16.0",
+    "babel-runtime": "6.20.0",
+    "chai": "4.0.0-canary.1",
+    "cross-env": "3.1.3",
+    "enzyme": "2.7.0",
+    "eslint": "3.12.2",
     "eslint-config-vazco": "2.2.0",
     "eslint-plugin-babel": "4.0.0",
-    "eslint-plugin-react": "6.7.1",
+    "eslint-plugin-react": "6.8.0",
     "eslint-plugin-vazco": "1.0.0",
-    "jsdom": "^9.8.3",
-    "material-ui": "^0.16.2",
-    "mocha": "^3.2.0",
-    "mock-require": "^2.0.0",
-    "nyc": "^10.0.0",
-    "react": "^15.0.0 || ^0.14.0",
-    "react-addons-test-utils": "^15.0.0 || ^0.14.0",
-    "react-dom": "^15.0.0 || ^0.14.0",
-    "rimraf": "^2.5.4",
-    "sinon": "^2.0.0-pre.4"
+    "jsdom": "9.9.1",
+    "material-ui": "0.16.5",
+    "mocha": "3.2.0",
+    "nyc": "10.0.0",
+    "react": "15.4.1",
+    "react-addons-test-utils": "15.4.1",
+    "react-dom": "15.4.1",
+    "rimraf": "2.5.4",
+    "sinon": "2.0.0-pre.4"
   },
   "peerDependencies": {
-    "material-ui": "^0.16.2",
+    "material-ui": "^0.16.5",
     "react": "^15.0.0 || ^0.14.0",
-    "uniforms": "^1.7.0"
+    "uniforms": "^1.8.1"
   },
   "babel": {
     "plugins": [
@@ -91,10 +90,7 @@
     "root": true,
     "extends": [
       "vazco"
-    ],
-    "env": {
-      "mocha": true
-    }
+    ]
   },
   "nyc": {
     "reporter": [
@@ -107,6 +103,6 @@
     "instrument": false
   },
   "dependencies": {
-    "babel-runtime": "^6.18.0"
+    "babel-runtime": "^6.20.0"
   }
 }
