+++ packages/uniforms-material/src/HiddenField.js
@@ -1,7 +1,7 @@
-import React            from 'react';
-import {BaseField}      from 'uniforms';
-import {filterDOMProps} from 'uniforms';
-import {nothing}        from 'uniforms';
+import BaseField      from 'uniforms/BaseField';
+import React          from 'react';
+import filterDOMProps from 'uniforms/filterDOMProps';
+import nothing        from 'uniforms/nothing';
 
 export default class HiddenField extends BaseField {
     componentWillReceiveProps ({value: valueDesired}) {
