+++ packages/uniforms-material/src/LongTextField.js
@@ -1,39 +1,33 @@
-import React            from 'react';
-import TextField        from 'material-ui/TextField';
-import {connectField}   from 'uniforms';
-import {filterDOMProps} from 'uniforms';
+import React          from 'react';
+import TextField      from 'material-ui/TextField';
+import connectField   from 'uniforms/connectField';
+import filterDOMProps from 'uniforms/filterDOMProps';
 
 const Text = ({
     disabled,
-    errorMessage,
     id,
+    errorMessage,
     inputRef,
     label,
     name,
     onChange,
     placeholder,
-    showInlineError,
-    type,
     value,
     ...props
 }) =>
     <TextField
         disabled={disabled}
-        errorText={showInlineError ? errorMessage : undefined}
+        errorText={errorMessage}
         floatingLabelText={label}
-        hintText={placeholder}
         id={id}
+        multiLine
         name={name}
         onChange={(event, value) => onChange(value)}
+        placeholder={placeholder}
         ref={inputRef}
-        type={type}
         value={value}
         {...filterDOMProps(props)}
     />
 ;
-
-Text.defaultProps = {
-    type: 'text'
-};
 
 export default connectField(Text);
