+++ packages/uniforms-material/src/index.js
@@ -1,24 +1,23 @@
-export {default as AutoField}     from './components/fields/AutoField';
-export {default as AutoFields}    from './components/fields/AutoFields';
-export {default as BoolField}     from './components/fields/BoolField';
-export {default as DateField}     from './components/fields/DateField';
-export {default as ErrorField}    from './components/fields/ErrorField';
-export {default as ErrorsField}   from './components/fields/ErrorsField';
-export {default as HiddenField}   from './components/fields/HiddenField';
-export {default as ListAddField}  from './components/fields/ListAddField';
-export {default as ListDelField}  from './components/fields/ListDelField';
-export {default as ListField}     from './components/fields/ListField';
-export {default as ListItemField} from './components/fields/ListItemField';
-export {default as LongTextField} from './components/fields/LongTextField';
-export {default as NestField}     from './components/fields/NestField';
-export {default as NumField}      from './components/fields/NumField';
-export {default as RadioField}    from './components/fields/RadioField';
-export {default as SelectField}   from './components/fields/SelectField';
-export {default as SubmitField}   from './components/fields/SubmitField';
-export {default as TextField}     from './components/fields/TextField';
-
-export {default as AutoForm}           from './components/forms/AutoForm';
-export {default as BaseForm}           from './components/forms/BaseForm';
-export {default as QuickForm}          from './components/forms/QuickForm';
-export {default as ValidatedForm}      from './components/forms/ValidatedForm';
-export {default as ValidatedQuickForm} from './components/forms/ValidatedQuickForm';
+export {default as AutoFields}         from './AutoFields';
+export {default as AutoField}          from './AutoField';
+export {default as AutoForm}           from './AutoForm';
+export {default as BaseForm}           from './BaseForm';
+export {default as BoolField}          from './BoolField';
+export {default as DateField}          from './DateField';
+export {default as ErrorField}         from './ErrorField';
+export {default as ErrorsField}        from './ErrorsField';
+export {default as HiddenField}        from './HiddenField';
+export {default as ListAddField}       from './ListAddField';
+export {default as ListDelField}       from './ListDelField';
+export {default as ListField}          from './ListField';
+export {default as ListItemField}      from './ListItemField';
+export {default as LongTextField}      from './LongTextField';
+export {default as NestField}          from './NestField';
+export {default as NumField}           from './NumField';
+export {default as QuickForm}          from './QuickForm';
+export {default as RadioField}         from './RadioField';
+export {default as SelectField}        from './SelectField';
+export {default as SubmitField}        from './SubmitField';
+export {default as TextField}          from './TextField';
+export {default as ValidatedForm}      from './ValidatedForm';
+export {default as ValidatedQuickForm} from './ValidatedQuickForm';
